+++ src/main/java/com/mercateo/common/rest/schemagen/RestJsonSchemaGenerator.java
@@ -18,6 +18,7 @@ import javax.ws.rs.core.Context;
 
 import com.mercateo.common.rest.schemagen.generator.JsonPropertyResult;
 import com.mercateo.common.rest.schemagen.generator.ObjectContextBuilder;
+import com.mercateo.common.rest.schemagen.json.mapper.PropertyJsonSchemaMapper;
 import org.glassfish.jersey.media.multipart.FormDataParam;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
